Add relations to posts, fix test, move all entities inside modules folder
