--- server/src/comment/entities/comment.entity.ts
@@ -1,34 +1,0 @@
-import { Post } from 'src/post/entities/post.entity';
-import { User } from 'src/user/entities/user.entity';
-import {
-  Column,
-  CreateDateColumn,
-  Entity,
-  JoinColumn,
-  ManyToOne,
-  PrimaryGeneratedColumn,
-  UpdateDateColumn,
-} from 'typeorm';
-
-@Entity('comments')
-export class Comment {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  message: string;
-
-  @CreateDateColumn()
-  created_at: Date;
-
-  @UpdateDateColumn()
-  updated_at: Date;
-
-  @ManyToOne(() => Post, (post) => post.comments)
-  @JoinColumn({ name: 'post_id' })
-  post: Post;
-
-  @ManyToOne(() => User, (user) => user.sentComments)
-  @JoinColumn({ name: 'sender_id' })
-  sender: User;
-}
