+++ server/src/db/seed/main.seeder.ts
@@ -1,5 +1,5 @@
-import { Post } from '../../post/entities/post.entity';
-import { User } from '../../user/entities/user.entity';
+import { Post } from '../../modules/post/entities/post.entity';
+import { User } from '../../modules/user/entities/user.entity';
 import { DataSource } from 'typeorm';
 import { Seeder, SeederFactoryManager } from 'typeorm-extension';
