--- server/src/messages/entities/message.entity.ts
@@ -1,45 +1,0 @@
-import { Chat } from 'src/chat/entities/chat.entity';
-import { User } from 'src/user/entities/user.entity';
-import {
-  Column,
-  CreateDateColumn,
-  Entity,
-  JoinColumn,
-  ManyToOne,
-  OneToMany,
-  PrimaryGeneratedColumn,
-  UpdateDateColumn,
-} from 'typeorm';
-import { MessageReaction } from './message_reaction.entity';
-
-@Entity('messages')
-export class Message {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  message: string;
-
-  @CreateDateColumn()
-  created_at: Date;
-
-  @UpdateDateColumn()
-  updated_at: Date;
-
-  @ManyToOne(() => Chat, (chat) => chat.messages, {
-    nullable: false,
-    onDelete: 'CASCADE',
-  })
-  @JoinColumn({ name: 'chat_id' })
-  chat: Chat;
-
-  @ManyToOne(() => User, (user) => user.sentMessages, {
-    nullable: false,
-    onDelete: 'CASCADE',
-  })
-  @JoinColumn({ name: 'sender_id' })
-  sender: User;
-
-  @OneToMany(() => MessageReaction, (reaction) => reaction.message)
-  reactions: MessageReaction[];
-}
