+++ server/src/modules/country/entities/country.entity.ts
@@ -1,0 +1,14 @@
+import { User } from 'src/modules/user/entities/user.entity';
+import { Column, Entity, OneToMany, PrimaryGeneratedColumn } from 'typeorm';
+
+@Entity('countries')
+export class Country {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column({ unique: true })
+  name: string;
+
+  @OneToMany(() => User, (user) => user.country)
+  users: User[];
+}
