+++ server/src/modules/post/entities/post.entity.ts
@@ -1,0 +1,48 @@
+import { PostTag } from 'src/modules/tag/entities/post_tag.entity';
+import { User } from '../../user/entities/user.entity';
+import { Comment } from 'src/modules/comment/entities/comment.entity';
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  JoinColumn,
+  ManyToOne,
+  OneToMany,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+} from 'typeorm';
+import { Chat } from 'src/modules/chat/entities/chat.entity';
+
+@Entity('posts')
+export class Post {
+  @PrimaryGeneratedColumn()
+  id: number;
+
+  @Column()
+  title: string;
+
+  @Column()
+  description: string;
+
+  @Column()
+  group_size: number;
+
+  @CreateDateColumn()
+  created_at: Date;
+
+  @UpdateDateColumn()
+  updated_at: Date;
+
+  @ManyToOne(() => User, (user) => user.posts)
+  @JoinColumn({ name: 'user_id' })
+  user: User;
+
+  @OneToMany(() => PostTag, (postTag) => postTag.post)
+  postTags: PostTag[];
+
+  @OneToMany(() => Comment, (comment) => comment.post)
+  comments: Comment[];
+
+  @OneToMany(() => Chat, (chat) => chat.post)
+  chats: Chat[];
+}
