+++ server/src/modules/post/post.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { PostController } from './post.controller';
+import { PostService } from './post.service';
+import { UserModule } from 'src/modules/user/user.module';
+import { Post } from './entities/post.entity';
+
+@Module({
+  imports: [TypeOrmModule.forFeature([Post]), UserModule],
+  controllers: [PostController],
+  providers: [PostService],
+})
+export class PostModule {}
