--- server/src/notification/entities/notification.entity.ts
@@ -1,31 +1,0 @@
-import { User } from 'src/user/entities/user.entity';
-import {
-  Column,
-  CreateDateColumn,
-  Entity,
-  JoinColumn,
-  ManyToOne,
-  PrimaryGeneratedColumn,
-} from 'typeorm';
-
-@Entity('notifications')
-export class Notification {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  type: string;
-
-  @Column()
-  content: string;
-
-  @Column({ default: false })
-  is_read: boolean;
-
-  @CreateDateColumn()
-  created_at: Date;
-
-  @ManyToOne(() => User, (user) => user.notifications)
-  @JoinColumn({ name: 'user_id' })
-  user: User;
-}
