--- server/src/post/post.module.ts
@@ -1,13 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { PostController } from './post.controller';
-import { PostService } from './post.service';
-import { UserModule } from 'src/user/user.module';
-import { Post } from './entities/post.entity';
-
-@Module({
-  imports: [TypeOrmModule.forFeature([Post]), UserModule],
-  controllers: [PostController],
-  providers: [PostService],
-})
-export class PostModule {}
