--- server/src/user/user.service.ts
@@ -1,76 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { User } from './entities/user.entity';
-import { Repository } from 'typeorm';
-import { RegisterUserDto } from './dtos/RegisterUser.dto';
-import { DuplicatedEmailException } from 'src/shared/exceptions/DuplictedEmail.exception';
-import bcrypt from 'bcrypt';
-import { CountryService } from 'src/country/country.service';
-
-enum Auth {
-  SALT_ROUNDS = 10,
-}
-
-@Injectable()
-export class UserService {
-  constructor(
-    @InjectRepository(User) private readonly usersRepository: Repository<User>,
-    private readonly countryService: CountryService,
-  ) {}
-
-  /**
-   * @param email is required.
-   * @returns an existing user account that matches provided email address, otherwise, returns null.
-   */
-  findByEmail(email: string): Promise<User | null> {
-    return this.usersRepository.findOne({
-      where: {
-        email,
-      },
-    });
-  }
-
-  /**
-   * @param id is required.
-   * @returns an existing user account that matches provided ID, otherwise, returns null.
-   */
-  findById(id: number): Promise<User | null> {
-    return this.usersRepository.findOne({
-      where: {
-        id,
-      },
-    });
-  }
-
-  async create(body: RegisterUserDto): Promise<User> {
-    const { username, gender, country, email, password } = body;
-
-    const duplicatedUser = await this.findByEmail(email);
-    if (duplicatedUser) {
-      throw new DuplicatedEmailException(email);
-    }
-
-    const countryEntity = await this.countryService.findOrCreateByName(country);
-
-    return this.usersRepository.save(
-      this.usersRepository.create({
-        username,
-        gender,
-        country: countryEntity,
-        email,
-        reputation: 0,
-        password: await this.hashPassword(password),
-        created_at: new Date(),
-        updated_at: new Date(),
-      }),
-    );
-  }
-
-  /**
-   * @param password is required.
-   * @returns a hashed password, generated using hardcoded salt rounds.
-   */
-  private hashPassword(password: string): Promise<string> {
-    return bcrypt.hash(password, Auth.SALT_ROUNDS);
-  }
-}
